finish dev / prod split
