+++ snowpack.config.dev.js
@@ -1,18 +1,15 @@
 module.exports = {
     plugins: [['@snowpack/plugin-typescript']],
-    root: 'src/example',
+    root: 'src',
     optimize: {
-        manifest: true,
-        minify: true,
         sourceMap: true,
-        splitting: true,
-        treeshake: true,
     },
     buildOptions: {
         out: 'dist',
     },
     mount: {
         dist: '/',
-        src: '/',
-    },
+        ['example']: '/',
+        ['src/lib']: '/lib',
+    }
 }
