--- src/example/index-old.ts
@@ -1,13 +1,0 @@
-// export * from './sketch'
-// import p5 from 'p5'
-// export { p5 }
-
-// import * as p5Global from 'p5/global'
-
-// declare global {
-//     interface p5 {}
-//     interface Window {
-//         setup: any
-//         draw: any
-//     }
-// }
