+++ src/lib/index.ts
@@ -1,3 +1,4 @@
+export * from './addFunction'
 export * from './Component'
 export * from './Sketch'
 export * from './Events'
